refactor: move replication rpc to dataserver
